arbitrage-ai-agent/arbitrage-bot/arbitrage_bot.py
import asyncio
import logging
from typing import Dict, List, Optional
import numpy as np
from datetime import datetime
import tensorflow as tf
from web3 import Web3
from near_api import Account as NearAccount
from zcash.core import ZcashTransaction
from .price_monitor import PriceMonitor
from .ml_predictor import MLPredictor
from .chain_interface import ChainInterface
from .privacy_manager import PrivacyManager
import json

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class ArbitrageBot:
    def __init__(self, config_path: str):
        """Initialize the arbitrage bot with configuration"""
        self.config = self._load_config(config_path)
        self.chain_interface = ChainInterface(self.config['chains'])
        self.ml_predictor = MLPredictor()
        self.privacy_manager = PrivacyManager()
        self.running = False
        self.trade_history = []
        self.total_profit = 0.0

    def _load_config(self, config_path: str) -> Dict:
        """Load configuration from file"""
        try:
            with open(config_path, 'r') as f:
                return json.load(f)
        except Exception as e:
            logger.error(f"Error loading config: {e}")
            raise

    async def start(self):
        """Start the arbitrage bot"""
        logger.info("Starting arbitrage bot...")
        self.running = True
        
        try:
            while self.running:
                # Monitor prices and find opportunities
                opportunities = await self._find_arbitrage_opportunities()
                
                if opportunities:
                    for opp in opportunities:
                        # Validate opportunity with ML predictions
                        if await self._validate_opportunity(opp):
                            # Execute trades with privacy protection
                            result = await self._execute_arbitrage(opp)
                            if result['success']:
                                self._update_trade_history(result)
                
                # Wait before next iteration
                await asyncio.sleep(self.config['monitoring']['interval'])
                
        except Exception as e:
            logger.error(f"Error in arbitrage bot main loop: {e}")
            self.stop()
            raise

    def stop(self):
        """Stop the arbitrage bot"""
        logger.info("Stopping arbitrage bot...")
        self.running = False

    async def _find_arbitrage_opportunities(self) -> List[Dict]:
        """Find arbitrage opportunities across chains"""
        try:
            opportunities = []
            chains = self.config['chains'].keys()
            
            # Get current prices across all chains
            prices = {}
            for chain in chains:
                token_address = self.config['chains'][chain]['zec_token_address']
                prices[chain] = await self.chain_interface.get_price(chain, token_address)
            
            # Find price differences that exceed minimum spread
            min_spread = self.config['trading']['min_spread']
            for buy_chain in chains:
                for sell_chain in chains:
                    if buy_chain != sell_chain:
                        spread = (prices[sell_chain] - prices[buy_chain]) / prices[buy_chain]
                        
                        if spread > min_spread:
                            opportunities.append({
                                'buy_chain': buy_chain,
                                'sell_chain': sell_chain,
                                'buy_price': prices[buy_chain],
                                'sell_price': prices[sell_chain],
                                'spread': spread,
                                'timestamp': datetime.now().isoformat()
                            })
            
            return opportunities
            
        except Exception as e:
            logger.error(f"Error finding arbitrage opportunities: {e}")
            return []

    async def _validate_opportunity(self, opportunity: Dict) -> bool:
        """Validate arbitrage opportunity using ML predictions"""
        try:
            # Get price predictions for both chains
            buy_prediction = await self.ml_predictor.predict_movements(
                chain=opportunity['buy_chain'],
                current_price=opportunity['buy_price']
            )
            
            sell_prediction = await self.ml_predictor.predict_movements(
                chain=opportunity['sell_chain'],
                current_price=opportunity['sell_price']
            )
            
            # Calculate expected profit considering predictions
            expected_spread = (
                sell_prediction['predicted_price'] - buy_prediction['predicted_price']
            ) / buy_prediction['predicted_price']
            
            # Check if opportunity remains profitable after fees
            total_fees = self._calculate_total_fees(opportunity)
            min_profit = self.config['trading']['min_profit']
            
            return expected_spread > (total_fees + min_profit)
            
        except Exception as e:
            logger.error(f"Error validating opportunity: {e}")
            return False

    async def _execute_arbitrage(self, opportunity: Dict) -> Dict:
        """Execute arbitrage trades with privacy protection"""
        try:
            amount = self._calculate_trade_amount(opportunity)
            
            # Prepare shielded transactions
            privacy_params = await self.privacy_manager.prepare_shielded_transaction(
                buy_chain=opportunity['buy_chain'],
                sell_chain=opportunity['sell_chain'],
                amount=amount
            )
            
            # Execute buy trade
            buy_result = await self.chain_interface.execute_trade(
                chain=opportunity['buy_chain'],
                is_buy=True,
                amount=amount,
                privacy_params=privacy_params['buy_params']
            )
            
            if buy_result['status'] == 'pending':
                # Wait for buy confirmation
                await self._wait_for_confirmation(
                    opportunity['buy_chain'],
                    buy_result['tx_hash']
                )
                
                # Execute sell trade
                sell_result = await self.chain_interface.execute_trade(
                    chain=opportunity['sell_chain'],
                    is_buy=False,
                    amount=amount,
                    privacy_params=privacy_params['sell_params']
                )
                
                if sell_result['status'] == 'pending':
                    # Wait for sell confirmation
                    await self._wait_for_confirmation(
                        opportunity['sell_chain'],
                        sell_result['tx_hash']
                    )
                    
                    # Calculate actual profit
                    profit = self._calculate_profit(
                        opportunity,
                        amount,
                        buy_result,
                        sell_result
                    )
                    
                    return {
                        'success': True,
                        'buy_tx': buy_result['tx_hash'],
                        'sell_tx': sell_result['tx_hash'],
                        'profit': profit,
                        'timestamp': datetime.now().isoformat(),
                        'opportunity': opportunity
                    }
            
            return {'success': False}
            
        except Exception as e:
            logger.error(f"Error executing arbitrage: {e}")
            return {'success': False}

    def _calculate_total_fees(self, opportunity: Dict) -> float:
        """Calculate total fees for both trades"""
        buy_fees = self.config['chains'][opportunity['buy_chain']]['fees']
        sell_fees = self.config['chains'][opportunity['sell_chain']]['fees']
        return buy_fees + sell_fees

    def _calculate_trade_amount(self, opportunity: Dict) -> float:
        """Calculate optimal trade amount based on opportunity"""
        max_amount = self.config['trading']['max_amount']
        min_amount = self.config['trading']['min_amount']
        
        # Calculate amount based on spread
        suggested_amount = min_amount + (
            (max_amount - min_amount) * (opportunity['spread'] / 100)
        )
        
        return min(suggested_amount, max_amount)

    async def _wait_for_confirmation(self, chain: str, tx_hash: str):
        """Wait for transaction confirmation"""
        max_attempts = self.config['monitoring']['max_confirmation_attempts']
        attempt = 0
        
        while attempt < max_attempts:
            # Check transaction status
            # Implementation would depend on chain-specific methods
            await asyncio.sleep(self.config['monitoring']['confirmation_interval'])
            attempt += 1

    def _calculate_profit(
        self,
        opportunity: Dict,
        amount: float,
        buy_result: Dict,
        sell_result: Dict
    ) -> float:
        """Calculate actual profit from completed trades"""
        # Implementation would calculate actual profit
        # considering executed prices and fees
        return amount * opportunity['spread']

    def _update_trade_history(self, result: Dict):
        """Update trade history and total profit"""
        self.trade_history.append(result)
        self.total_profit += result['profit']
        
        # Save trade history to file
        self._save_trade_history()

    def _save_trade_history(self):
        """Save trade history to file"""
        try:
            with open(self.config['logging']['trade_history_file'], 'w') as f:
                json.dump(self.trade_history, f, indent=2)
        except Exception as e:
            logger.error(f"Error saving trade history: {e}")

    def get_stats(self) -> Dict:
        """Get current bot statistics"""
        return {
            'total_profit': self.total_profit,
            'total_trades': len(self.trade_history),
            'success_rate': self._calculate_success_rate(),
            'average_profit': self._calculate_average_profit()
        }

    def _calculate_success_rate(self) -> float:
        """Calculate success rate of trades"""
        if not self.trade_history:
            return 0.0
        successful_trades = len([t for t in self.trade_history if t['success']])
        return successful_trades / len(self.trade_history)

    def _calculate_average_profit(self) -> float:
        """Calculate average profit per trade"""
        if not self.trade_history:
            return 0.0
        return self.total_profit / len(self.trade_history) 