arbitrage-ai-agent/arbitrage-bot/chain_interface.py
from typing import Dict, List, Optional
import logging
import asyncio
from web3 import Web3
from solana.rpc.async_api import AsyncClient as SolanaClient
from bitcoinrpc.authproxy import AuthServiceProxy

logger = logging.getLogger(__name__)

class ChainInterface:
    def __init__(self, config: Dict):
        """Initialize chain connections and configurations"""
        self.config = config
        self.connections = {}
        self.setup_connections()

    def setup_connections(self):
        """Set up connections to different blockchain networks"""
        try:
            # Setup Ethereum connection
            if 'eth' in self.config:
                self.connections['ETH'] = Web3(Web3.HTTPProvider(
                    self.config['eth']['rpc_url']
                ))

            # Setup Solana connection
            if 'sol' in self.config:
                self.connections['SOL'] = SolanaClient(
                    self.config['sol']['rpc_url']
                )

            # Setup Bitcoin connection
            if 'btc' in self.config:
                self.connections['BTC'] = AuthServiceProxy(
                    f"http://{self.config['btc']['user']}:{self.config['btc']['password']}@"
                    f"{self.config['btc']['host']}:{self.config['btc']['port']}"
                )

        except Exception as e:
            logger.error(f"Error setting up chain connections: {e}")
            raise

    async def get_price(self, chain: str, token_address: str) -> float:
        """Get current price of ZEC on specified chain"""
        try:
            if chain == 'ETH':
                return await self._get_eth_price(token_address)
            elif chain == 'SOL':
                return await self._get_sol_price(token_address)
            elif chain == 'BTC':
                return await self._get_btc_price()
            else:
                raise ValueError(f"Unsupported chain: {chain}")

        except Exception as e:
            logger.error(f"Error getting price for {chain}: {e}")
            raise

    async def execute_trade(
        self,
        chain: str,
        is_buy: bool,
        amount: float,
        privacy_params: Dict
    ) -> Dict:
        """Execute a trade on specified chain with privacy features"""
        try:
            if chain == 'ETH':
                return await self._execute_eth_trade(is_buy, amount, privacy_params)
            elif chain == 'SOL':
                return await self._execute_sol_trade(is_buy, amount, privacy_params)
            elif chain == 'BTC':
                return await self._execute_btc_trade(is_buy, amount, privacy_params)
            else:
                raise ValueError(f"Unsupported chain: {chain}")

        except Exception as e:
            logger.error(f"Error executing trade on {chain}: {e}")
            raise

    async def _get_eth_price(self, token_address: str) -> float:
        """Get ZEC price from Ethereum DEX"""
        w3 = self.connections['ETH']
        # Implementation would use DEX contract to get price
        # This is a placeholder
        return 100.0

    async def _get_sol_price(self, token_address: str) -> float:
        """Get ZEC price from Solana DEX"""
        client = self.connections['SOL']
        # Implementation would use Serum/Orca API to get price
        # This is a placeholder
        return 100.0

    async def _get_btc_price(self) -> float:
        """Get ZEC price from Bitcoin exchanges"""
        btc_client = self.connections['BTC']
        # Implementation would use exchange API to get price
        # This is a placeholder
        return 100.0

    async def _execute_eth_trade(
        self,
        is_buy: bool,
        amount: float,
        privacy_params: Dict
    ) -> Dict:
        """Execute trade on Ethereum"""
        w3 = self.connections['ETH']
        try:
            # Build transaction with privacy parameters
            tx = {
                'from': privacy_params['note']['pool_address'],
                'value': w3.to_wei(amount, 'ether'),
                'gas': privacy_params['note']['chain_specific_params']['gas_limit'],
                'gasPrice': await self._get_eth_gas_price(),
                'nonce': await self._get_eth_nonce(privacy_params['note']['pool_address']),
                'data': self._build_eth_swap_data(is_buy, amount)
            }

            # Sign and send transaction
            signed_tx = w3.eth.account.sign_transaction(tx, privacy_params['proof'])
            tx_hash = w3.eth.send_raw_transaction(signed_tx.rawTransaction)

            return {
                'tx_hash': tx_hash.hex(),
                'status': 'pending'
            }

        except Exception as e:
            logger.error(f"Error executing Ethereum trade: {e}")
            raise

    async def _execute_sol_trade(
        self,
        is_buy: bool,
        amount: float,
        privacy_params: Dict
    ) -> Dict:
        """Execute trade on Solana"""
        client = self.connections['SOL']
        try:
            # Build transaction with privacy parameters
            tx = {
                'recent_blockhash': await self._get_sol_blockhash(),
                'fee_payer': privacy_params['note']['pool_address'],
                'instructions': self._build_sol_swap_instructions(is_buy, amount)
            }

            # Sign and send transaction
            signed_tx = await self._sign_sol_transaction(tx, privacy_params['proof'])
            tx_hash = await client.send_transaction(signed_tx)

            return {
                'tx_hash': tx_hash['result'],
                'status': 'pending'
            }

        except Exception as e:
            logger.error(f"Error executing Solana trade: {e}")
            raise

    async def _execute_btc_trade(
        self,
        is_buy: bool,
        amount: float,
        privacy_params: Dict
    ) -> Dict:
        """Execute trade on Bitcoin"""
        btc_client = self.connections['BTC']
        try:
            # Build transaction with privacy parameters
            tx = {
                'version': privacy_params['note']['chain_specific_params']['version'],
                'locktime': privacy_params['note']['chain_specific_params']['locktime'],
                'vin': self._build_btc_inputs(amount),
                'vout': self._build_btc_outputs(is_buy, amount, privacy_params['note']['pool_address'])
            }

            # Sign and send transaction
            signed_tx = btc_client.signrawtransactionwithkey(
                tx,
                [privacy_params['proof']]
            )
            tx_hash = btc_client.sendrawtransaction(signed_tx['hex'])

            return {
                'tx_hash': tx_hash,
                'status': 'pending'
            }

        except Exception as e:
            logger.error(f"Error executing Bitcoin trade: {e}")
            raise

    async def _get_eth_gas_price(self) -> int:
        """Get current Ethereum gas price"""
        w3 = self.connections['ETH']
        return w3.eth.gas_price

    async def _get_eth_nonce(self, address: str) -> int:
        """Get next nonce for Ethereum address"""
        w3 = self.connections['ETH']
        return w3.eth.get_transaction_count(address)

    def _build_eth_swap_data(self, is_buy: bool, amount: float) -> bytes:
        """Build Ethereum swap transaction data"""
        # Implementation would encode DEX swap function call
        # This is a placeholder
        return b''

    async def _get_sol_blockhash(self) -> str:
        """Get recent Solana blockhash"""
        client = self.connections['SOL']
        response = await client.get_recent_blockhash()
        return response['result']['value']['blockhash']

    def _build_sol_swap_instructions(self, is_buy: bool, amount: float) -> List:
        """Build Solana swap instructions"""
        # Implementation would create Serum/Orca swap instructions
        # This is a placeholder
        return []

    async def _sign_sol_transaction(self, tx: Dict, proof: bytes) -> Dict:
        """Sign Solana transaction"""
        # Implementation would sign transaction with proof
        # This is a placeholder
        return tx

    def _build_btc_inputs(self, amount: float) -> List:
        """Build Bitcoin transaction inputs"""
        # Implementation would select appropriate UTXOs
        # This is a placeholder
        return []

    def _build_btc_outputs(
        self,
        is_buy: bool,
        amount: float,
        address: str
    ) -> List:
        """Build Bitcoin transaction outputs"""
        # Implementation would create appropriate outputs
        # This is a placeholder
        return [] 