requirements.txt
# Core dependencies
web3==6.11.1
solana==0.30.1
python-bitcoinrpc==1.0
aiohttp==3.9.1
asyncio==3.4.3

# Machine Learning
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2

# Privacy and Cryptography
pycryptodome==3.19.0
zcash==0.2.0
pyblake2==1.1.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
python-telegram-bot==20.7
schedule==1.2.1

# Monitoring and Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0 